new 27_8

+++ src/app/update_file.py
@@ -1,12 +1,13 @@
-import streamlit as st
-import shutil
-import patoolib
-import os
 import base64
+import os
 import secrets
-import pandas as pd
-from src.db.funtion_database import offline_edit
+import shutil
 import subprocess
+
+import pandas as pd
+import streamlit as st
+
+from src.db.funtion_database import offline_edit
 
 
 @st.cache_resource
@@ -61,19 +62,32 @@ def update_file_into_server_new(car_name, List_file, csrf_token):
             if '仕様表' in file.name and (".zip" in file.name or ".7z" in file.name):
                 extract(file_path, folder_save_file_upload)
                 os.remove(file_path)
-        elif car_name != "" and car_name.upper() in file.name.upper():
-            car_name = str(car_name).upper()
-            folder_save_file_upload = os.path.join('./data', car_name)
-            if not os.path.exists(folder_save_file_upload):
-                os.makedirs(folder_save_file_upload)
-            file_path = os.path.join(folder_save_file_upload, file.name)
-            with open(file_path, "wb") as f:
-                f.write(file.getvalue())
+            st.session_state.message_10 = "仕様表 Updated"
 
-            # ===================extract file .zip, .7z=======================
-            if car_name in file.name and (".zip" in file.name or ".7z" in file.name):
-                extract(file_path, folder_save_file_upload)
-                os.remove(file_path)
+        else:
+            if car_name != "" and car_name.upper() in file.name.upper():
+                car_name = str(car_name).upper()
+                folder_save_file_upload = os.path.join('./data', car_name)
+                if not os.path.exists(folder_save_file_upload):
+                    os.makedirs(folder_save_file_upload)
+                file_path = os.path.join(folder_save_file_upload, file.name)
+                with open(file_path, "wb") as f:
+                    f.write(file.getvalue())
+
+                # ===================extract file .zip, .7z=======================
+                if car_name in file.name and (".zip" in file.name or ".7z" in file.name):
+                    extract(file_path, folder_save_file_upload)
+                    os.remove(file_path)
+                syo_files = [f for f in os.listdir(folder_save_file_upload) if f.endswith('.xlsx') and f.startswith('仕様表')]
+                if len(syo_files) > 0:
+                    destination_syo_folder = os.path.join('./data', "仕様表")
+                    os.makedirs(destination_syo_folder, exist_ok=True)
+                    for item in syo_files:
+                        source_item_1 = os.path.join(folder_save_file_upload, item)
+                        shutil.copy2(source_item_1, destination_syo_folder)
+                st.session_state.message_10 = "プロ管 Updated"
+            elif car_name == "" or car_name.upper() not in file.name.upper():
+                st.session_state.message_10 = 'Check model code or file name'
 
 
 def update_file_after_edit(code, pwt, plant, case, file_updates, csrf_token, name_user):
@@ -101,44 +115,46 @@ def update_file_after_edit(code, pwt, plant, case, file_updates, csrf_token, nam
         return ""
 
 
-# def extract(file_path, output_dir):
-#     # Tạo thư mục đích nếu nó chưa tồn tại
-#     if not os.path.exists(output_dir):
-#         os.makedirs(output_dir)
-#     seven_zip_path = shutil.which("7z")
-#
-#     # Nếu không tìm thấy 7z trong PATH, sử dụng đường dẫn mặc định
-#     if seven_zip_path is None:
-#         seven_zip_path = "C:\\Program Files\\7-Zip\\7z.exe"  # Thay thế bằng đường dẫn đầy đủ đến 7z.exe nếu cần
-#
-#     if not os.path.exists(seven_zip_path):
-#         raise FileNotFoundError(
-#             f"The command '{seven_zip_path}' was not found. Please install it or add it to your PATH.")
-#
-#     if ".zip" in file_path:
-#         subprocess.run([seven_zip_path, "x", file_path, f"-o{output_dir}", "-y"])
-#         folder = file_path[:-4]  # Lấy đường dẫn mà không có phần mở rộng .zip
-#
-#     if ".7z" in file_path:
-#         subprocess.run([seven_zip_path, "x", file_path, f"-o{output_dir}", "-y"])
-#         folder = file_path[:-3]  # Lấy đường dẫn mà không có phần mở rộng .7z
-#     move_all_items(folder, output_dir)
-#     delete_folder(folder)
 def extract(file_path, output_dir):
     # Tạo thư mục đích nếu nó chưa tồn tại
     if not os.path.exists(output_dir):
         os.makedirs(output_dir)
-    # Giải nén tệp .7z vào thư mục đích
-    # subprocess.run(["7z","x",file_path,f"-o{output_dir}","-mcp=UTF-8"])
-    # patoolib.extract_archive(file_path, outdir=output_dir)
+    seven_zip_path = shutil.which("7z")
+
+    # Nếu không tìm thấy 7z trong PATH, sử dụng đường dẫn mặc định
+    if seven_zip_path is None:
+        seven_zip_path = "C:\\Program Files\\7-Zip\\7z.exe"  # Thay thế bằng đường dẫn đầy đủ đến 7z.exe nếu cần
+
+    if not os.path.exists(seven_zip_path):
+        raise FileNotFoundError(
+            f"The command '{seven_zip_path}' was not found. Please install it or add it to your PATH.")
+
     if ".zip" in file_path:
-        subprocess.run(["unar", file_path, "-o", output_dir])
-        folder = file_path[0:len(file_path) - 4]
+        subprocess.run([seven_zip_path, "x", file_path, f"-o{output_dir}", "-y"])
+        folder = file_path[:-4]  # Lấy đường dẫn mà không có phần mở rộng .zip
+
     if ".7z" in file_path:
-        subprocess.run(["7z", "x", file_path, f"-o{output_dir}"])
-        folder = file_path[0:len(file_path) - 3]
+        subprocess.run([seven_zip_path, "x", file_path, f"-o{output_dir}", "-y"])
+        folder = file_path[:-3]  # Lấy đường dẫn mà không có phần mở rộng .7z
     move_all_items(folder, output_dir)
     delete_folder(folder)
+
+
+# def extract(file_path, output_dir):
+#     # Tạo thư mục đích nếu nó chưa tồn tại
+#     if not os.path.exists(output_dir):
+#         os.makedirs(output_dir)
+#     # Giải nén tệp .7z vào thư mục đích
+#     # subprocess.run(["7z","x",file_path,f"-o{output_dir}","-mcp=UTF-8"])
+#     # patoolib.extract_archive(file_path, outdir=output_dir)
+#     if ".zip" in file_path:
+#         subprocess.run(["unar", file_path, "-o", output_dir])
+#         folder = file_path[0:len(file_path) - 4]
+#     if ".7z" in file_path:
+#         subprocess.run(["7z", "x", file_path, f"-o{output_dir}"])
+#         folder = file_path[0:len(file_path) - 3]
+#     move_all_items(folder, output_dir)
+#     delete_folder(folder)
 
 
 def move_all_items(source_dir, destination_dir):
@@ -157,6 +173,6 @@ def move_all_items(source_dir, destination_dir):
 def delete_folder(folder_path):
     if os.path.exists(folder_path):
         shutil.rmtree(folder_path)
-        print(f'Thư mục {folder_path} đã được xóa.')
+        # print(f'Thư mục {folder_path} đã được xóa.')
     else:
         print(f'Thư mục {folder_path} không tồn tại.')
